+++ .eslintrc.js
@@ -1,8 +1,6 @@
-'use strict';
-
 module.exports = {
-  'root': true,
-  'plugins': [
+  root: true,
+  plugins: [
     'jsdoc',
     'mocha',
     'promise',
@@ -13,28 +11,33 @@ module.exports = {
   extends: [
     'eslint:recommended',
     'airbnb-base',
-    'plugin:import/typescript',
-    'plugin:@typescript-eslint/eslint-recommended',
-    'plugin:@typescript-eslint/recommended',
-    'plugin:@typescript-eslint/recommended-requiring-type-checking',
   ],
-  'env': {
-    'node': true,
-    'es6': true
+  env: {
+    node: true,
+    es6: true,
   },
-  'settings': {
+  settings: {
     'import/resolver': {
-      'node': {
-        'extensions': ['.js', '.ts']
-      }
-    }
+      node: {
+        extensions: ['.js', '.ts'],
+      },
+    },
+    jsdoc: {
+      preferredTypes: {
+        Array: 'Array<object>',
+        'Array.': 'Array<object>',
+        'Array<>': '[]',
+        'Array.<>': '[]',
+        'Promise.<>': 'Promise<>',
+      },
+    },
   },
   rules: {
-    'curly': ['error', 'all'],
+    curly: ['error', 'all'],
     'callback-return': ['error', ['callback', 'cb', 'next', 'done']],
     'class-methods-use-this': 'off',
     'consistent-return': 'off',
-    'handle-callback-err': ['error', '^.*err' ],
+    'handle-callback-err': ['error', '^.*err'],
     'new-cap': 'off',
     'no-console': 'error',
     'no-else-return': 'error',
@@ -50,32 +53,26 @@ module.exports = {
     'no-unsafe-negation': 'error',
     'no-use-before-define': ['error', 'nofunc'],
     'no-useless-rename': 'error',
-    'padding-line-between-statements': ['error',
-      { 'blankLine': 'always', 'prev': [
-          'directive',
-          'block',
-          'block-like',
-          'multiline-block-like',
-          'cjs-export',
-          'cjs-import',
-          'class',
-          'export',
-          'import',
-          'if'
-        ], 'next': '*' },
-      { 'blankLine': 'never', 'prev': 'directive', 'next': 'directive' },
-      { 'blankLine': 'any', 'prev': '*', 'next': ['if', 'for', 'cjs-import', 'import'] },
-      { 'blankLine': 'any', 'prev': ['export', 'import'], 'next': ['export', 'import'] },
-      { 'blankLine': 'always', 'prev': '*', 'next': ['try', 'function', 'switch'] },
-      { 'blankLine': 'always', 'prev': 'if', 'next': 'if' },
-      { 'blankLine': 'never', 'prev': ['return', 'throw'], 'next': '*' }
+    'padding-line-between-statements': [
+      'error',
+      {
+        blankLine: 'always',
+        prev: ['directive', 'block', 'block-like', 'multiline-block-like', 'cjs-export', 'cjs-import', 'class', 'export', 'import', 'if'],
+        next: '*',
+      },
+      { blankLine: 'never', prev: 'directive', next: 'directive' },
+      { blankLine: 'any', prev: '*', next: ['if', 'for', 'cjs-import', 'import'] },
+      { blankLine: 'any', prev: ['export', 'import'], next: ['export', 'import'] },
+      { blankLine: 'always', prev: '*', next: ['try', 'function', 'switch'] },
+      { blankLine: 'always', prev: 'if', next: 'if' },
+      { blankLine: 'never', prev: ['return', 'throw'], next: '*' },
     ],
-    'strict': ['error', 'safe'],
+    strict: ['error', 'safe'],
     'no-new': 'off',
     'no-empty': 'error',
     'no-empty-function': 'error',
     'valid-jsdoc': 'off',
-    'yoda': 'error',
+    yoda: 'error',
 
     'import/extensions': ['error', 'never'],
     'import/no-unresolved': 'off',
@@ -101,7 +98,7 @@ module.exports = {
 
     'promise/always-return': 'error',
     'promise/always-catch': 'off',
-    'promise/catch-or-return': ['error', {'allowThen': true }],
+    'promise/catch-or-return': ['error', { allowThen: true }],
     'promise/no-native': 'off',
     'promise/param-names': 'error',
 
@@ -120,59 +117,49 @@ module.exports = {
     'security/detect-unsafe-regex': 'error',
 
     // Override airbnb
-    'eqeqeq': ['error', 'smart'],
+    eqeqeq: ['error', 'smart'],
     'func-names': 'error',
-    'id-length': ['error', {'exceptions': ['_', '$', 'e', 'i', 'j', 'k', 'q', 'x', 'y']}],
+    'id-length': ['error', { exceptions: ['_', '$', 'e', 'i', 'j', 'k', 'q', 'x', 'y'] }],
     'no-param-reassign': 'off', // Work toward enforcing this rule
-    'radix': 'off',
+    radix: 'off',
     'spaced-comment': 'off',
     'max-len': 'off',
     'no-continue': 'off',
     'no-plusplus': 'off',
     'no-prototype-builtins': 'off',
-    'no-restricted-syntax': [
+    'no-restricted-syntax': ['error', 'DebuggerStatement', 'LabeledStatement', 'WithStatement'],
+    'no-restricted-properties': [
       'error',
-      'DebuggerStatement',
-      'LabeledStatement',
-      'WithStatement'
+      {
+        object: 'arguments',
+        property: 'callee',
+        message: 'arguments.callee is deprecated',
+      },
+      {
+        property: '__defineGetter__',
+        message: 'Please use Object.defineProperty instead.',
+      },
+      {
+        property: '__defineSetter__',
+        message: 'Please use Object.defineProperty instead.',
+      },
     ],
-    'no-restricted-properties': ['error', {
-      'object': 'arguments',
-      'property': 'callee',
-      'message': 'arguments.callee is deprecated'
-    }, {
-      'property': '__defineGetter__',
-      'message': 'Please use Object.defineProperty instead.'
-    }, {
-      'property': '__defineSetter__',
-      'message': 'Please use Object.defineProperty instead.'
-    }],
     'no-useless-escape': 'off',
-    'object-shorthand': ['error', 'always', {
-      'ignoreConstructors': false,
-      'avoidQuotes': true,
-      'avoidExplicitReturnArrows': true
-    }],
+    'object-shorthand': [
+      'error',
+      'always',
+      {
+        ignoreConstructors: false,
+        avoidQuotes: true,
+        avoidExplicitReturnArrows: true,
+      },
+    ],
     // 'prefer-arrow-callback': ['error', { 'allowNamedFunctions': true }],
     'prefer-spread': 'error',
-    'prefer-destructuring': 'off'
+    'prefer-destructuring': 'off',
   },
   overrides: [{
-    files: [
-      '**/*.tests.ts',
-    ],
-    rules: {
-      'mocha/no-async-describe': 'error',
-      'mocha/no-exclusive-tests': 'error',
-      'mocha/no-global-tests': 'error',
-      'mocha/no-identical-title': 'error',
-      'mocha/no-nested-tests': 'error',
-      'mocha/no-pending-tests': 'error',
-    }
-  }, {
-    files: [
-      '*.ts',
-    ],
+    files: ['*.ts'],
     parser: '@typescript-eslint/parser',
     parserOptions: {
       project: './tsconfig.lint.json',
@@ -180,12 +167,14 @@ module.exports = {
     extends: [
       'eslint:recommended',
       'airbnb-base',
-      'plugin:@typescript-eslint/recommended',
       'plugin:import/typescript',
+      'plugin:@typescript-eslint/eslint-recommended',
+      'plugin:@typescript-eslint/recommended',
+      'plugin:@typescript-eslint/recommended-requiring-type-checking',
     ],
     rules: {
       'class-methods-use-this': 'off',
-      'indent': 'off',
+      indent: 'off',
       'max-len': 'off',
       'no-dupe-class-members': 'off',
       'no-extra-semi': 'off',
@@ -194,12 +183,8 @@ module.exports = {
       'no-underscore-dangle': 'off',
       'no-useless-constructor': 'off',
       'no-unused-expressions': 'error',
-      'no-restricted-syntax': [
-        'error',
-        'DebuggerStatement',
-        'LabeledStatement',
-        'WithStatement'
-      ],
+      'no-restricted-syntax': ['error', 'DebuggerStatement', 'LabeledStatement', 'WithStatement'],
+      'no-use-before-define': 'off',
 
       'import/prefer-default-export': 'off',
       'import/no-cycle': 'off',
@@ -209,32 +194,33 @@ module.exports = {
       '@typescript-eslint/array-type': ['error', { default: 'array' }],
       '@typescript-eslint/await-thenable': 'error',
       '@typescript-eslint/adjacent-overload-signatures': 'error',
-      '@typescript-eslint/class-name-casing': 'error',
       '@typescript-eslint/consistent-type-assertions': 'error',
       '@typescript-eslint/consistent-type-definitions': 'error',
       '@typescript-eslint/no-extraneous-class': 'error',
       '@typescript-eslint/explicit-function-return-type': 'error',
       '@typescript-eslint/explicit-member-accessibility': ['error'],
-      '@typescript-eslint/generic-type-naming': 'error',
-      '@typescript-eslint/interface-name-prefix': ['error', 'never'],
-      '@typescript-eslint/member-ordering': ['error', {
-        default: [
-          // Index signature
-          'signature',
-          // Fields
-          'private-field',
-          'public-field',
-          'protected-field',
-          // Constructors
-          'public-constructor',
-          'protected-constructor',
-          'private-constructor',
-          // Methods
-          'public-method',
-          'protected-method',
-          'private-method',
-        ],
-      }],
+      '@typescript-eslint/naming-convention': 'error',
+      '@typescript-eslint/member-ordering': [
+        'error',
+        {
+          default: [
+            // Index signature
+            'signature',
+            // Fields
+            'private-field',
+            'public-field',
+            'protected-field',
+            // Constructors
+            'public-constructor',
+            'protected-constructor',
+            'private-constructor',
+            // Methods
+            'public-method',
+            'protected-method',
+            'private-method',
+          ],
+        },
+      ],
       '@typescript-eslint/no-array-constructor': 'error',
       '@typescript-eslint/no-empty-interface': 'error',
       '@typescript-eslint/no-explicit-any': 'error',
@@ -261,5 +247,19 @@ module.exports = {
       '@typescript-eslint/unbound-method': 'error',
       '@typescript-eslint/unified-signatures': 'error',
     },
-  }]
+  }, {
+    files: ['**/*.tests.ts'],
+    rules: {
+      '@typescript-eslint/no-unsafe-member-access': 'off',
+      '@typescript-eslint/no-unsafe-call': 'off',
+      '@typescript-eslint/restrict-template-expressions': 'off',
+
+      'mocha/no-async-describe': 'error',
+      'mocha/no-exclusive-tests': 'error',
+      'mocha/no-global-tests': 'error',
+      'mocha/no-identical-title': 'error',
+      'mocha/no-nested-tests': 'error',
+      'mocha/no-pending-tests': 'error',
+    },
+  }],
 };