+++ src/Entity.ts
@@ -1,3 +1,5 @@
+export type EntityFieldValue = string | string[] | number | number[] | Date | boolean | boolean[] | Record<string, unknown> | null | unknown;
+
 export interface Entity {
   // eslint-disable-next-line @typescript-eslint/no-explicit-any
   [index: string]: any;