+++ src/decorators/versionColumn.ts
@@ -6,11 +6,13 @@ import {
 } from '../metadata';
 import { ColumnTypeOptions } from './ColumnTypeOptions';
 
+// eslint-disable-next-line @typescript-eslint/ban-types
 type ReturnFunctionType = (object: object, propertyName: string) => void;
 
 export function versionColumn(options?: ColumnTypeOptions): ReturnFunctionType;
 export function versionColumn(dbColumnName: string, options?: ColumnTypeOptions): ReturnFunctionType;
 export function versionColumn(dbColumnNameOrOptions?: string | ColumnTypeOptions, options?: ColumnTypeOptions): ReturnFunctionType {
+  // eslint-disable-next-line @typescript-eslint/ban-types
   return function versionColumnDecorator(object: object, propertyName: string): void {
     let dbColumnName: string | undefined;
     if (typeof dbColumnNameOrOptions === 'string') {