+++ src/query/FindOneResult.ts
@@ -6,5 +6,5 @@ import { ChainablePromiseLike } from '../ChainablePromiseLike';
 export interface FindOneResult<TEntity extends Entity> extends ChainablePromiseLike<TEntity | null> {
   where(args: WhereQuery): FindOneResult<TEntity>;
   populate(propertyName: Extract<keyof TEntity, string>, options?: PopulateArgs): FindOneResult<TEntity>;
-  sort(value: string | object): FindOneResult<TEntity>;
+  sort(value: string | Record<string, number | string>): FindOneResult<TEntity>;
 }