+++ src/query/FindResult.ts
@@ -4,7 +4,7 @@ import { ChainablePromiseLike } from '../ChainablePromiseLike';
 
 export interface FindResult<TEntity> extends ChainablePromiseLike<TEntity[]> {
   where(args: WhereQuery): FindResult<TEntity>;
-  sort(value: string | object): FindResult<TEntity>;
+  sort(value: string | Record<string, number | string>): FindResult<TEntity>;
   limit(value: number): FindResult<TEntity>;
   skip(value: number): FindResult<TEntity>;
   paginate(options: PaginateOptions): FindResult<TEntity>;