+++ src/query/WhereQuery.ts
@@ -1,6 +1,6 @@
-import { Entity } from '../Entity';
+import { Entity, EntityFieldValue } from '../Entity';
 
-export type WhereClauseValue = string | string[] | number | number[] | Date | boolean | object | null | Entity | WhereQuery;
+export type WhereClauseValue = EntityFieldValue | Entity | WhereQuery;
 
 export interface WhereQuery {
   [index: string]: WhereClauseValue;